apps/vitastate-backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
pypdf
